third_party/WebKit/LayoutTests/css3/device-adapt/opera/constrain-015-expected.txt


FAIL CSS Test: @viewport constrained - zoom less than min-zoom. assert_equals: expected 400 but got 800
